Commit Six
A “back to top” button was added
Back to top button slides like a tab on bottom right
It will slide in when the user scrolls down past 55 pixels
It slides out and hides when the user scrolls up past 55 pixels
Back to top button has a fade in and out effect, exactly like the navbar
Back to top button works and brings the page to the very top
Small bug where if the navbar is hidden and the user clicks the back to
top button, the navbar will stay hidden. Need to fix this, but it is a
minor problem and I will sort it out later.